app2/requirements.txt
streamlit
python-dotenv
google.generativeai
langchain
langchain-google-genai 
